refactor(gateway): GAR-479 — Q10.e extract build_whatsapp_channels to bootstrap/whatsapp.rs
Move build_whatsapp_channels (~180 LOC) from bootstrap/mod.rs into a new
sibling module bootstrap/whatsapp.rs. Re-exported at the same public path
via `pub use whatsapp::build_whatsapp_channels;` — zero behaviour change.

mod.rs shrinks ~1878 → ~1698 LOC. Returns Vec<Arc<WhatsAppChannel>>
(distinct from Slack/Discord's Vec<Box<dyn Channel>>).
